Moved ParserContext to its own file so that its fields will be private within the parser submodule
